requirements.txt
streamlit
google-genai
sentence-transformers
faiss-cpu
pypdf
langchain-text-splitters
requests
python-dotenv
